Conversion to NSInteger - 64 bit warning fix.

+++ MKAdditions/MKBlockAdditions.h
@@ -8,7 +8,7 @@
 
 typedef void (^VoidBlock)();
 
-typedef void (^DismissBlock)(int buttonIndex);
+typedef void (^DismissBlock)(NSInteger buttonIndex);
 typedef void (^CancelBlock)();
 typedef void (^PhotoPickedBlock)(UIImage *chosenImage);
